refactor integration tests and disable ones that use chain
