+++ .gitlab-ci.yml
@@ -59,59 +59,59 @@ test_cargo:
     after_script:
     - *cleanup_obsolete_cache
 
-# Regression tests.
-test_regression:
-    stage: rust
-    before_script:
-    - *setup_rust
-    - *setup_cache
-    - *install_protos
-    script:
-    - rustc --version && cargo --version
-    - ./scripts/build_wasm.sh
-    - cargo test --package nearcore --test test_tps_regression test --features regression_tests
-    tags:
-        - regression_tests
-    after_script:
-    - *cleanup_obsolete_cache
-    only:
-        - schedules
-
-# A set of expensive tests.
-test_cases_testnet_direct:
-    stage: rust
-    before_script:
-    - *setup_rust
-    - *setup_cache
-    - *install_protos
-    script:
-    - rustc --version && cargo --version
-    - ./scripts/build_wasm.sh
-    - cargo test --package nearcore --test test_cases_testnet_direct test --features expensive_tests
-    tags:
-      - expensive_tests
-    after_script:
-    - *cleanup_obsolete_cache
-    only:
-        - schedules
-
-# A set of expensive tests.
-test_cases_testnet_rpc:
-    stage: rust
-    before_script:
-    - *setup_rust
-    - *setup_cache
-    - *install_protos
-    script:
-    - rustc --version && cargo --version
-    - ./scripts/build_wasm.sh
-    - cargo test --package nearcore --test test_cases_testnet_rpc test --features expensive_tests
-    tags:
-        - expensive_tests
-    after_script:
-    - *cleanup_obsolete_cache
-    only:
-        - schedules
+## Regression tests.
+#test_regression:
+#    stage: rust
+#    before_script:
+#    - *setup_rust
+#    - *setup_cache
+#    - *install_protos
+#    script:
+#    - rustc --version && cargo --version
+#    - ./scripts/build_wasm.sh
+#    - cargo test --package nearcore --test test_tps_regression test --features regression_tests
+#    tags:
+#        - regression_tests
+#    after_script:
+#    - *cleanup_obsolete_cache
+#    only:
+#        - schedules
+#
+## A set of expensive tests.
+#test_cases_testnet_direct:
+#    stage: rust
+#    before_script:
+#    - *setup_rust
+#    - *setup_cache
+#    - *install_protos
+#    script:
+#    - rustc --version && cargo --version
+#    - ./scripts/build_wasm.sh
+#    - cargo test --package nearcore --test test_cases_testnet_direct test --features expensive_tests
+#    tags:
+#      - expensive_tests
+#    after_script:
+#    - *cleanup_obsolete_cache
+#    only:
+#        - schedules
+#
+## A set of expensive tests.
+#test_cases_testnet_rpc:
+#    stage: rust
+#    before_script:
+#    - *setup_rust
+#    - *setup_cache
+#    - *install_protos
+#    script:
+#    - rustc --version && cargo --version
+#    - ./scripts/build_wasm.sh
+#    - cargo test --package nearcore --test test_cases_testnet_rpc test --features expensive_tests
+#    tags:
+#        - expensive_tests
+#    after_script:
+#    - *cleanup_obsolete_cache
+#    only:
+#        - schedules
 
 release_build:
     stage: rust
