+++ tests/test_custom_handler.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "old_tests")]
 #[cfg(test)]
 mod test {
     use std::sync::Arc;
