+++ tests/test_simple.rs
@@ -1,4 +1,5 @@
 ///! Simply starts and runs TestNet for a while.
+#[cfg(feature = "old_tests")]
 #[cfg(test)]
 mod test {
     use std::sync::Arc;
